ëdocs: comment final final

+++ mookeun.sql
@@ -1,7 +1,7 @@
 CREATE TABLE comments(
-    id int auto_increment primary key,
-    post_id int not null,
-    parent_id int,
+    id bigint auto_increment primary key,
+    post_id bigint not null,
+    parent_id bigint,
     contents varchar(200),
     created_at timestamp default current_timestamp,
     modified_at timestamp default current_timestamp on update current_timestamp,
@@ -10,9 +10,9 @@ CREATE TABLE comments(
 )ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
 
 create table comment_likes(
-    id int auto_increment primary key,
-    user_id int not null,
-    post_id int not null,
+    id bigint auto_increment primary key,
+    user_id bigint not null,
+    post_id bigint not null,
     created_at timestamp default current_timestamp,
     foreign key (user_id) references users(user_id) on delete cascade,
     foreign key (post_id) references posts(id) on delete cascade
